requirements.txt
aniso8601==9.0.1
astroid==2.15.0
attrs==22.2.0
bcrypt==4.0.1
blinker==1.5
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
dill==0.3.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Principal==0.4.0
flask-restx==1.1.0
Flask-Security-Too==5.1.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
passlib==1.7.4
platformdirs==3.1.0
pycparser==2.21
pylint==2.16.3
PyMySQL==1.0.2
pyrsistent==0.19.3
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.5.post1
SQLAlchemy-serializer==1.4.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
